refactor(integrationtests): unify the two Soroban executors
executeSorobanOperationAs was a near-verbatim copy of
executeSorobanOperation that fixed two real defects only in the copy:
it preserved simulation-assigned auth nonces (nonce 0 is one-shot per
address) and re-simulated after signing so MinResourceFee reflects the
signed entries' size. One executeSorobanOperation(op, source,
extraSigners, retries) now carries both behaviors for every caller.

Sequence resolution branches on the source: the master account drives
SharedContainers' locally tracked counter, any other actor's sequence is
fetched from RPC. Since every master submission now advances the local
counter, the As-path drift SyncMasterSequence existed to repair can no
longer occur, so it and its re-sync call sites are removed.

Auth entries an available keypair cannot sign now fail loudly naming the
required address (the old shared path silently left them unsigned);
source-account-credentialed entries pass through unchanged as before.
