+++ internal/integrationtests/infrastructure/blend_operations.go
@@ -1,31 +1,18 @@
 // Package infrastructure provides Soroban transaction helpers for integration tests.
 //
-// This file adds two things needed to exercise the Blend v2 pools protocol:
-//   - pure ScVal builder functions for the Blend/SEP-40 contract UDTs, and
-//   - a Soroban executor that drives a transaction from an arbitrary actor keypair, rather than
-//     the shared master account executeSorobanOperation always uses.
+// This file adds the pure ScVal builder functions for the Blend/SEP-40 contract UDTs;
+// transactions built from them are driven by executeSorobanOperation
+// (soroban_transactions.go), which takes the acting keypair as its source parameter.
 package infrastructure
 
 import (
-	"bytes"
-	"context"
 	"encoding/binary"
-	"encoding/json"
 	"fmt"
 	"math/big"
-	"net/http"
-	"strconv"
 	"testing"
 
-	"github.com/stellar/go-stellar-sdk/keypair"
-	"github.com/stellar/go-stellar-sdk/strkey"
-	"github.com/stellar/go-stellar-sdk/txnbuild"
 	"github.com/stellar/go-stellar-sdk/xdr"
 	"github.com/stretchr/testify/require"
-
-	"github.com/stellar/wallet-backend/internal/entities"
-	"github.com/stellar/wallet-backend/internal/utils"
-	"github.com/stellar/wallet-backend/pkg/sorobanauth"
 )
 
 // BlendRequest mirrors the Blend v2 pool Request struct. RequestType encodes the pool action:
@@ -295,307 +282,4 @@ func scSep40StellarAssetVec(t *testing.T, addrs []string) xdr.ScVal {
 		vals = append(vals, scSep40Asset(t, a))
 	}
 	return scVec(vals...)
-}
-
-// scAddressToString converts an xdr.ScAddress back into its G- or C-address string form. It is
-// the inverse of parseAddressToScAddress, used to name the required signer when an auth entry's
-// address credentials don't match any available keypair.
-func scAddressToString(addr xdr.ScAddress) (string, error) {
-	switch addr.Type {
-	case xdr.ScAddressTypeScAddressTypeAccount:
-		if addr.AccountId == nil {
-			return "", fmt.Errorf("account address has a nil AccountId")
-		}
-		return addr.AccountId.Address(), nil
-	case xdr.ScAddressTypeScAddressTypeContract:
-		if addr.ContractId == nil {
-			return "", fmt.Errorf("contract address has a nil ContractId")
-		}
-		encoded, err := strkey.Encode(strkey.VersionByteContract, addr.ContractId[:])
-		if err != nil {
-			return "", fmt.Errorf("encoding contract address: %w", err)
-		}
-		return encoded, nil
-	default:
-		return "", fmt.Errorf("unsupported ScAddress type %d", addr.Type)
-	}
-}
-
-// signAuthEntriesAs signs simulation-returned Soroban authorization entries using whichever
-// keypair among signers matches each entry's required address. Source-account-credentialed
-// entries need no explicit signature (the transaction signature covers them) and are passed
-// through unchanged. Unlike signAuthEntries (which always signs with a single fixed keypair and
-// hardcodes nonce 0, which is only safe for the master account's first-ever Soroban call), this
-// preserves the nonce simulation assigned to each entry and fails loudly, naming the required
-// address, if no provided keypair can sign it.
-func signAuthEntriesAs(
-	authEntries []xdr.SorobanAuthorizationEntry,
-	signers []*keypair.Full,
-	latestLedger int64,
-) ([]xdr.SorobanAuthorizationEntry, error) {
-	if len(authEntries) == 0 {
-		return authEntries, nil
-	}
-
-	authSigner := sorobanauth.AuthSigner{NetworkPassphrase: networkPassphrase}
-	signed := make([]xdr.SorobanAuthorizationEntry, len(authEntries))
-	for i, entry := range authEntries {
-		switch entry.Credentials.Type {
-		case xdr.SorobanCredentialsTypeSorobanCredentialsSourceAccount:
-			signed[i] = entry
-		case xdr.SorobanCredentialsTypeSorobanCredentialsAddress:
-			requiredAddress, err := scAddressToString(entry.Credentials.Address.Address)
-			if err != nil {
-				return nil, fmt.Errorf("resolving required auth address for entry %d: %w", i, err)
-			}
-
-			var matched *keypair.Full
-			for _, signer := range signers {
-				if signer.Address() == requiredAddress {
-					matched = signer
-					break
-				}
-			}
-			if matched == nil {
-				return nil, fmt.Errorf("no signer available for required auth address %s (entry %d)", requiredAddress, i)
-			}
-
-			nonce := int64(entry.Credentials.Address.Nonce)
-			signedEntry, err := authSigner.AuthorizeEntry(entry, nonce, uint32(latestLedger+LedgerValidityBuffer), matched)
-			if err != nil {
-				return nil, fmt.Errorf("signing auth entry %d for %s: %w", i, requiredAddress, err)
-			}
-			signed[i] = signedEntry
-		default:
-			signed[i] = entry
-		}
-	}
-	return signed, nil
-}
-
-// getAccountSequenceRPC fetches an account's current ledger sequence number directly via RPC.
-// executeSorobanOperationAs uses this (rather than a locally tracked counter) because, unlike the
-// shared master account, arbitrary actor keypairs are not tracked by SharedContainers.
-func getAccountSequenceRPC(client *http.Client, rpcURL, address string) (int64, error) {
-	keyXDR, err := utils.GetAccountLedgerKey(address)
-	if err != nil {
-		return 0, fmt.Errorf("building account ledger key: %w", err)
-	}
-
-	requestBody := map[string]interface{}{
-		"jsonrpc": "2.0",
-		"id":      1,
-		"method":  "getLedgerEntries",
-		"params": map[string][]string{
-			"keys": {keyXDR},
-		},
-	}
-	jsonBody, err := json.Marshal(requestBody)
-	if err != nil {
-		return 0, fmt.Errorf("marshaling request: %w", err)
-	}
-
-	resp, err := client.Post(rpcURL, "application/json", bytes.NewBuffer(jsonBody))
-	if err != nil {
-		return 0, fmt.Errorf("posting to RPC: %w", err)
-	}
-	defer func() {
-		_ = resp.Body.Close() //nolint:errcheck
-	}()
-
-	var rpcResp struct {
-		Result entities.RPCGetLedgerEntriesResult `json:"result"`
-		Error  *struct {
-			Code    int    `json:"code"`
-			Message string `json:"message"`
-		} `json:"error"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&rpcResp); err != nil {
-		return 0, fmt.Errorf("decoding response: %w", err)
-	}
-	if rpcResp.Error != nil {
-		return 0, fmt.Errorf("RPC error: %s", rpcResp.Error.Message)
-	}
-	if len(rpcResp.Result.Entries) == 0 {
-		return 0, fmt.Errorf("no ledger entry found for account %s", address)
-	}
-
-	var data xdr.LedgerEntryData
-	if err := xdr.SafeUnmarshalBase64(rpcResp.Result.Entries[0].DataXDR, &data); err != nil {
-		return 0, fmt.Errorf("decoding account ledger entry: %w", err)
-	}
-	account, ok := data.GetAccount()
-	if !ok {
-		return 0, fmt.Errorf("ledger entry for %s is not an account", address)
-	}
-	return int64(account.SeqNum), nil
-}
-
-// executeSorobanOperationAs executes an InvokeHostFunction with source as the transaction source
-// account and primary auth signer (any additional required signers are passed via extraSigners).
-//
-// It mirrors the 11-step pattern of executeSorobanOperation, adapted for an arbitrary actor:
-//  1. Fetch source's CURRENT sequence via RPC (no locally tracked counter — only the shared
-//     master account gets one, since SharedContainers doesn't own arbitrary actor keypairs).
-//  2. Build and simulate the transaction.
-//  3. Sign simulation-returned auth entries: source-account-credentialed entries need no
-//     signature; address-credentialed entries are matched to a keypair in [source]+extraSigners
-//     and signed preserving the nonce simulation assigned (never hardcoded to 0). If no keypair
-//     matches, fail with an error naming the required address.
-//  4. Re-simulate with the signed auth entries so the resource footprint (and MinResourceFee)
-//     reflects their actual signed size, mirroring the double-simulate pattern in
-//     Fixtures.prepareSimulateAndSignContractOp.
-//  5. Apply the final simulation's SorobanData and MinResourceFee, rebuild incrementing source's
-//     sequence, sign with source, submit, and wait for confirmation.
-//
-// Every current Blend wrapper call site (blend_contracts.go) passes nil extraSigners and
-// DefaultConfirmationRetries, since none of the pinned Blend contract calls need a second signer
-// or a non-default retry budget; both params stay general for callers that do.
-//
-//nolint:unparam // see above: uniform today, kept general for future callers
-func (s *SharedContainers) executeSorobanOperationAs(
-	ctx context.Context,
-	t *testing.T,
-	op *txnbuild.InvokeHostFunction,
-	source *keypair.Full,
-	extraSigners []*keypair.Full,
-	retries int,
-) (string, error) {
-	rpcURL, err := s.RPCContainer.GetConnectionString(ctx)
-	if err != nil {
-		return "", fmt.Errorf("getting RPC connection string: %w", err)
-	}
-
-	// Step 1: fetch source's current sequence from RPC.
-	seq, err := getAccountSequenceRPC(s.httpClient, rpcURL, source.Address())
-	if err != nil {
-		return "", fmt.Errorf("getting source account sequence: %w", err)
-	}
-	sourceAccount := &txnbuild.SimpleAccount{AccountID: source.Address(), Sequence: seq}
-
-	signers := make([]*keypair.Full, 0, 1+len(extraSigners))
-	signers = append(signers, source)
-	signers = append(signers, extraSigners...)
-
-	// Step 2: build and simulate.
-	tx, err := txnbuild.NewTransaction(txnbuild.TransactionParams{
-		SourceAccount:        sourceAccount,
-		Operations:           []txnbuild.Operation{op},
-		BaseFee:              txnbuild.MinBaseFee,
-		IncrementSequenceNum: false,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewTimeout(DefaultTransactionTimeout),
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("building transaction for simulation: %w", err)
-	}
-	txXDR, err := tx.Base64()
-	if err != nil {
-		return "", fmt.Errorf("encoding transaction for simulation: %w", err)
-	}
-	simulationResult, err := simulateTransactionRPC(s.httpClient, rpcURL, txXDR)
-	if err != nil {
-		return "", fmt.Errorf("simulating transaction: %w", err)
-	}
-	if simulationResult.Error != "" {
-		return "", fmt.Errorf("simulation failed: %s", simulationResult.Error)
-	}
-
-	// Step 3+4: sign auth entries (preserving the simulation nonce) and re-simulate so the
-	// resource footprint accounts for the signed entries' size.
-	if len(simulationResult.Results) > 0 && len(simulationResult.Results[0].Auth) > 0 {
-		signedAuth, err := signAuthEntriesAs(simulationResult.Results[0].Auth, signers, simulationResult.LatestLedger)
-		if err != nil {
-			return "", err
-		}
-		op.Auth = signedAuth
-
-		tx, err = txnbuild.NewTransaction(txnbuild.TransactionParams{
-			SourceAccount:        sourceAccount,
-			Operations:           []txnbuild.Operation{op},
-			BaseFee:              txnbuild.MinBaseFee,
-			IncrementSequenceNum: false,
-			Preconditions: txnbuild.Preconditions{
-				TimeBounds: txnbuild.NewTimeout(DefaultTransactionTimeout),
-			},
-		})
-		if err != nil {
-			return "", fmt.Errorf("rebuilding transaction with signed auth: %w", err)
-		}
-		txXDR, err = tx.Base64()
-		if err != nil {
-			return "", fmt.Errorf("encoding transaction with signed auth: %w", err)
-		}
-		simulationResult, err = simulateTransactionRPC(s.httpClient, rpcURL, txXDR)
-		if err != nil {
-			return "", fmt.Errorf("re-simulating transaction: %w", err)
-		}
-		if simulationResult.Error != "" {
-			return "", fmt.Errorf("re-simulation failed: %s", simulationResult.Error)
-		}
-	}
-
-	// Step 5: apply resources, rebuild for real submission, sign, submit, wait for confirmation.
-	op.Ext = xdr.TransactionExt{
-		V:           1,
-		SorobanData: &simulationResult.TransactionData,
-	}
-
-	minResourceFee, err := strconv.ParseInt(simulationResult.MinResourceFee, 10, 64)
-	if err != nil {
-		return "", fmt.Errorf("parsing MinResourceFee: %w", err)
-	}
-
-	tx, err = txnbuild.NewTransaction(txnbuild.TransactionParams{
-		SourceAccount:        sourceAccount,
-		Operations:           []txnbuild.Operation{op},
-		BaseFee:              minResourceFee + txnbuild.MinBaseFee,
-		IncrementSequenceNum: true,
-		Preconditions: txnbuild.Preconditions{
-			TimeBounds: txnbuild.NewTimeout(DefaultTransactionTimeout),
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("rebuilding transaction: %w", err)
-	}
-
-	tx, err = tx.Sign(networkPassphrase, source)
-	if err != nil {
-		return "", fmt.Errorf("signing transaction: %w", err)
-	}
-	txXDR, err = tx.Base64()
-	if err != nil {
-		return "", fmt.Errorf("encoding signed transaction: %w", err)
-	}
-
-	sendResult, err := submitTransactionToRPC(s.httpClient, rpcURL, txXDR)
-	if err != nil {
-		return "", fmt.Errorf("submitting transaction: %w", err)
-	}
-	if sendResult.Status == entities.ErrorStatus {
-		return "", fmt.Errorf("transaction failed with status: %s, hash: %s, errorResultXdr: %s",
-			sendResult.Status, sendResult.Hash, sendResult.ErrorResultXDR)
-	}
-
-	if err := waitForTransactionConfirmation(ctx, t, s.httpClient, rpcURL, sendResult.Hash, retries); err != nil {
-		return "", fmt.Errorf("waiting for confirmation: %w", err)
-	}
-
-	return sendResult.Hash, nil
-}
-
-// SyncMasterSequence refreshes the master account's local sequence counter from RPC. Call it
-// before resuming master-account operations (executeSorobanOperation, executeClassicOperation) if
-// the master account may have submitted transactions through a path that doesn't track
-// SharedContainers' local counter — e.g. acting as an extraSigner in executeSorobanOperationAs —
-// so the counter doesn't drift from the ledger's view of the account's sequence number.
-func (s *SharedContainers) SyncMasterSequence(ctx context.Context, t *testing.T) {
-	rpcURL, err := s.RPCContainer.GetConnectionString(ctx)
-	require.NoError(t, err, "getting RPC connection string")
-
-	seq, err := getAccountSequenceRPC(s.httpClient, rpcURL, s.masterKeyPair.Address())
-	require.NoError(t, err, "fetching master account sequence")
-
-	s.masterAccount.Sequence = seq
 }
